\name{k.mode}
\alias{k.mode}
\title{Kernel Density Mode}
\description{
Function to return the mode (via kernel density) from numeric data.
}
\usage{
k.mode(data)
}
\arguments{
\item{data}{A vector, matrix, or array containing numerical values for analysis.}
}
\value{
Returns a numeric vector of length one.
}
